refactor(recruitment): split rest-controller god-object into 4 domain controllers (S2 of #141)
Pure code movement — no runtime behavior changes. The single
RecruitmentRestController (1456 LOC) is replaced by four focused
domain controllers + a shared support trait.

New controllers (each with its own register_routes()):

  - RecruitmentNoticesRestController        — 179 LOC,  3 endpoints
    (notices CRUD: list, create, update)
  - RecruitmentClassificationsRestController — 509 LOC, 9 endpoints
    (list, import, promote-preview, call, bulk-call, change status,
     change preview status, delete, cancel-call)
  - RecruitmentAdjutanciesRestController     — 420 LOC, 10 endpoints
    (adjutancy CRUD, notice attach/detach, reasons CRUD — Reasons
     folded here since both are catalog domains and the issue spec
     names only 4 controllers)
  - RecruitmentCandidatesRestController      — 344 LOC, 5 endpoints
    (candidates CRUD + me/recruitment)

Shared trait `RecruitmentRestSupport` (164 LOC) holds the 6 cap-check
methods and the 2 wp_error envelope helpers used by all controllers.

The loader (`RecruitmentLoader`) now instantiates and registers all
four controllers on `rest_api_init`. The original
`class-ffc-recruitment-rest-controller.php` is deleted (no
class_alias — user explicitly approved no aliases for this refactor).

Tests: `tests/Unit/RecruitmentRestControllerTest.php` updated to
instantiate the new controllers — cap-check tests use
RecruitmentNoticesRestController (any controller works since they
all `use` the trait); the route-registration tests aggregate
register_routes() from all four into a single accumulator and run
the original spot-checks unchanged (the union of registered routes
is identical to before).

Verification:
  - php -l: clean on all 5 new files + loader + test.
  - vendor/bin/phpunit --filter Recruitment: 227/227 OK.
  - vendor/bin/phpstan analyse includes/recruitment/ --level=8: no errors.
