--- includes/recruitment/class-ffc-recruitment-rest-controller.php
@@ -1,1456 +1,0 @@
-<?php
-/**
- * Recruitment REST Controller
- *
- * Single entry point for every admin REST route under the
- * `ffcertificate/v1/recruitment` namespace plus the candidate-self
- * `/me/recruitment` endpoint. Every callback is a thin dispatcher: it
- * sanitizes/validates input, delegates to the corresponding service or
- * repository, and returns a `\WP_REST_Response` or `\WP_Error`.
- *
- * Authorization model:
- *
- *   - Admin endpoints: `permission_callback` checks
- *     `current_user_can('ffc_manage_recruitment')`. The cap is granted to
- *     the `administrator` role on activation (sprint 3) and to the
- *     dedicated `ffc_recruitment_manager` role.
- *   - Candidate-self endpoint (`GET /me/recruitment`): just
- *     `is_user_logged_in()`. The query joins `candidate.user_id =
- *     current_user_id` so users only ever see their own data.
- *   - Public endpoints: NONE. The public shortcode (sprint 11) renders
- *     server-side; deliberately no public REST so we avoid the
- *     enumeration vector.
- *
- * Error contract: every failure returns a `\WP_Error` with a stable
- * `recruitment_*` code, an HTTP status (400/403/404/409/500), and an
- * optional `data` payload (e.g. `blocked_by` reference counts from the
- * delete service).
- *
- * @package FreeFormCertificate\Recruitment
- * @since   6.0.0
- */
-
-declare(strict_types=1);
-
-namespace FreeFormCertificate\Recruitment;
-
-use FreeFormCertificate\Core\DocumentFormatter;
-use FreeFormCertificate\Core\Encryption;
-use FreeFormCertificate\Core\SensitiveFieldRegistry;
-
-if ( ! defined( 'ABSPATH' ) ) {
-	exit;
-}
-
-/**
- * REST controller for the recruitment module.
- *
- * @phpstan-import-type CandidateRow from RecruitmentCandidateRepository
- */
-final class RecruitmentRestController {
-
-	/** Namespace for all admin + candidate-self routes. */
-	private const NAMESPACE = 'ffcertificate/v1';
-
-	/** Resource base. */
-	private const REST_BASE = 'recruitment';
-
-	/** Admin capability gating every admin route. */
-	private const ADMIN_CAP = 'ffc_manage_recruitment';
-
-	/**
-	 * Hook callback for `rest_api_init`.
-	 *
-	 * @return void
-	 */
-	public function register_routes(): void {
-		$ns   = self::NAMESPACE;
-		$base = '/' . self::REST_BASE;
-
-		// ── Notices ─────────────────────────────────────────────────────
-		register_rest_route(
-			$ns,
-			$base . '/notices',
-			array(
-				array(
-					'methods'             => \WP_REST_Server::READABLE,
-					'callback'            => array( $this, 'list_notices' ),
-					'permission_callback' => array( $this, 'check_admin_cap' ),
-					'args'                => array(
-						'status' => array(
-							'type'              => 'string',
-							'sanitize_callback' => 'sanitize_text_field',
-						),
-					),
-				),
-				array(
-					'methods'             => \WP_REST_Server::CREATABLE,
-					'callback'            => array( $this, 'create_notice' ),
-					'permission_callback' => array( $this, 'check_admin_cap' ),
-					'args'                => array(
-						'code' => array(
-							'type'              => 'string',
-							'required'          => true,
-							'sanitize_callback' => 'sanitize_text_field',
-						),
-						'name' => array(
-							'type'              => 'string',
-							'required'          => true,
-							'sanitize_callback' => 'sanitize_text_field',
-						),
-					),
-				),
-			)
-		);
-
-		register_rest_route(
-			$ns,
-			$base . '/notices/(?P<id>\d+)',
-			array(
-				'methods'             => \WP_REST_Server::EDITABLE,
-				'callback'            => array( $this, 'update_notice' ),
-				'permission_callback' => array( $this, 'check_admin_cap' ),
-			)
-		);
-
-		register_rest_route(
-			$ns,
-			$base . '/notices/(?P<id>\d+)/classifications',
-			array(
-				'methods'             => \WP_REST_Server::READABLE,
-				'callback'            => array( $this, 'list_classifications' ),
-				'permission_callback' => array( $this, 'check_admin_cap' ),
-				'args'                => array(
-					'list_type'    => array(
-						'type'              => 'string',
-						'sanitize_callback' => 'sanitize_text_field',
-					),
-					'adjutancy_id' => array(
-						'type'              => 'integer',
-						'sanitize_callback' => 'absint',
-					),
-				),
-			)
-		);
-
-		register_rest_route(
-			$ns,
-			$base . '/notices/(?P<id>\d+)/import',
-			array(
-				'methods'             => \WP_REST_Server::CREATABLE,
-				'callback'            => array( $this, 'import_csv' ),
-				'permission_callback' => array( $this, 'check_can_import_csv' ),
-			)
-		);
-
-		register_rest_route(
-			$ns,
-			$base . '/notices/(?P<id>\d+)/promote-preview',
-			array(
-				'methods'             => \WP_REST_Server::CREATABLE,
-				'callback'            => array( $this, 'promote_preview' ),
-				'permission_callback' => array( $this, 'check_can_import_csv' ),
-				'args'                => array(
-					'mode' => array(
-						'type'              => 'string',
-						'required'          => true,
-						'enum'              => array( 'snapshot', 'definitive_import' ),
-						'sanitize_callback' => 'sanitize_text_field',
-					),
-				),
-			)
-		);
-
-		// ── Classifications ─────────────────────────────────────────────
-		register_rest_route(
-			$ns,
-			$base . '/classifications/(?P<id>\d+)/call',
-			array(
-				'methods'             => \WP_REST_Server::CREATABLE,
-				'callback'            => array( $this, 'call_classification' ),
-				'permission_callback' => array( $this, 'check_can_call_candidates' ),
-				'args'                => array(
-					'date_to_assume'      => array(
-						'type'              => 'string',
-						'required'          => true,
-						'sanitize_callback' => 'sanitize_text_field',
-					),
-					'time_to_assume'      => array(
-						'type'              => 'string',
-						'required'          => true,
-						'sanitize_callback' => 'sanitize_text_field',
-					),
-					'out_of_order_reason' => array(
-						'type'              => 'string',
-						'sanitize_callback' => 'sanitize_text_field',
-					),
-					'notes'               => array(
-						'type'              => 'string',
-						'sanitize_callback' => 'sanitize_text_field',
-					),
-				),
-			)
-		);
-
-		register_rest_route(
-			$ns,
-			$base . '/classifications/bulk-call',
-			array(
-				'methods'             => \WP_REST_Server::CREATABLE,
-				'callback'            => array( $this, 'bulk_call_classifications' ),
-				'permission_callback' => array( $this, 'check_can_call_candidates' ),
-			)
-		);
-
-		register_rest_route(
-			$ns,
-			$base . '/classifications/(?P<id>\d+)/status',
-			array(
-				'methods'             => \WP_REST_Server::EDITABLE,
-				'callback'            => array( $this, 'change_classification_status' ),
-				'permission_callback' => array( $this, 'check_admin_cap' ),
-				'args'                => array(
-					'status' => array(
-						'type'              => 'string',
-						'required'          => true,
-						'sanitize_callback' => 'sanitize_text_field',
-					),
-					'reason' => array(
-						'type'              => 'string',
-						'sanitize_callback' => 'sanitize_text_field',
-					),
-				),
-			)
-		);
-
-		register_rest_route(
-			$ns,
-			$base . '/classifications/(?P<id>\d+)/preview-status',
-			array(
-				'methods'             => \WP_REST_Server::EDITABLE,
-				'callback'            => array( $this, 'change_classification_preview_status' ),
-				'permission_callback' => array( $this, 'check_admin_cap' ),
-				'args'                => array(
-					'preview_status'    => array(
-						'type'              => 'string',
-						'required'          => true,
-						'sanitize_callback' => 'sanitize_text_field',
-					),
-					'preview_reason_id' => array(
-						'type'              => 'integer',
-						'sanitize_callback' => 'absint',
-					),
-				),
-			)
-		);
-
-		register_rest_route(
-			$ns,
-			$base . '/classifications/(?P<id>\d+)',
-			array(
-				'methods'             => \WP_REST_Server::DELETABLE,
-				'callback'            => array( $this, 'delete_classification' ),
-				'permission_callback' => array( $this, 'check_admin_cap' ),
-			)
-		);
-
-		register_rest_route(
-			$ns,
-			$base . '/classifications/(?P<id>\d+)/calls/(?P<call_id>\d+)/cancel',
-			array(
-				'methods'             => \WP_REST_Server::CREATABLE,
-				'callback'            => array( $this, 'cancel_call' ),
-				'permission_callback' => array( $this, 'check_admin_cap' ),
-				'args'                => array(
-					'reason' => array(
-						'type'              => 'string',
-						'required'          => true,
-						'sanitize_callback' => 'sanitize_text_field',
-					),
-				),
-			)
-		);
-
-		// ── Adjutancies ─────────────────────────────────────────────────
-		register_rest_route(
-			$ns,
-			$base . '/adjutancies',
-			array(
-				array(
-					'methods'             => \WP_REST_Server::READABLE,
-					'callback'            => array( $this, 'list_adjutancies' ),
-					'permission_callback' => array( $this, 'check_admin_cap' ),
-				),
-				array(
-					'methods'             => \WP_REST_Server::CREATABLE,
-					'callback'            => array( $this, 'create_adjutancy' ),
-					'permission_callback' => array( $this, 'check_admin_cap' ),
-					'args'                => array(
-						'slug'  => array(
-							'type'              => 'string',
-							'required'          => true,
-							'sanitize_callback' => 'sanitize_title',
-						),
-						'name'  => array(
-							'type'              => 'string',
-							'required'          => true,
-							'sanitize_callback' => 'sanitize_text_field',
-						),
-						'color' => array(
-							'type'              => 'string',
-							'required'          => false,
-							'sanitize_callback' => 'sanitize_text_field',
-						),
-					),
-				),
-			)
-		);
-
-		register_rest_route(
-			$ns,
-			$base . '/adjutancies/(?P<id>\d+)',
-			array(
-				array(
-					'methods'             => \WP_REST_Server::EDITABLE,
-					'callback'            => array( $this, 'update_adjutancy' ),
-					'permission_callback' => array( $this, 'check_admin_cap' ),
-				),
-				array(
-					'methods'             => \WP_REST_Server::DELETABLE,
-					'callback'            => array( $this, 'delete_adjutancy' ),
-					'permission_callback' => array( $this, 'check_admin_cap' ),
-				),
-			)
-		);
-
-		// ── Reasons ─────────────────────────────────────────────────────
-		register_rest_route(
-			$ns,
-			$base . '/reasons',
-			array(
-				array(
-					'methods'             => \WP_REST_Server::READABLE,
-					'callback'            => array( $this, 'list_reasons' ),
-					'permission_callback' => array( $this, 'check_can_manage_reasons' ),
-				),
-				array(
-					'methods'             => \WP_REST_Server::CREATABLE,
-					'callback'            => array( $this, 'create_reason' ),
-					'permission_callback' => array( $this, 'check_can_manage_reasons' ),
-					'args'                => array(
-						'slug'       => array(
-							'type'              => 'string',
-							'required'          => true,
-							'sanitize_callback' => 'sanitize_title',
-						),
-						'label'      => array(
-							'type'              => 'string',
-							'required'          => true,
-							'sanitize_callback' => 'sanitize_text_field',
-						),
-						'color'      => array(
-							'type'              => 'string',
-							'required'          => false,
-							'sanitize_callback' => 'sanitize_text_field',
-						),
-						'applies_to' => array(
-							'type'     => 'array',
-							'required' => false,
-						),
-					),
-				),
-			)
-		);
-
-		register_rest_route(
-			$ns,
-			$base . '/reasons/(?P<id>\d+)',
-			array(
-				array(
-					'methods'             => \WP_REST_Server::EDITABLE,
-					'callback'            => array( $this, 'update_reason' ),
-					'permission_callback' => array( $this, 'check_can_manage_reasons' ),
-				),
-				array(
-					'methods'             => \WP_REST_Server::DELETABLE,
-					'callback'            => array( $this, 'delete_reason' ),
-					'permission_callback' => array( $this, 'check_can_manage_reasons' ),
-				),
-			)
-		);
-
-		// Notice ↔ Adjutancy attachment (N:N junction).
-		// PUT  /notices/{id}/adjutancies/{adjutancy_id} — attach
-		// DELETE /notices/{id}/adjutancies/{adjutancy_id} — detach
-		// The CSV importer requires every adjutancy referenced in the file
-		// to be attached to the target notice via this junction; without
-		// this route, admins had no UI/API path to perform the attachment.
-		register_rest_route(
-			$ns,
-			$base . '/notices/(?P<id>\d+)/adjutancies/(?P<adjutancy_id>\d+)',
-			array(
-				array(
-					'methods'             => \WP_REST_Server::EDITABLE,
-					'callback'            => array( $this, 'attach_notice_adjutancy' ),
-					'permission_callback' => array( $this, 'check_admin_cap' ),
-				),
-				array(
-					'methods'             => \WP_REST_Server::DELETABLE,
-					'callback'            => array( $this, 'detach_notice_adjutancy' ),
-					'permission_callback' => array( $this, 'check_admin_cap' ),
-				),
-			)
-		);
-
-		// ── Candidates ──────────────────────────────────────────────────
-		register_rest_route(
-			$ns,
-			$base . '/candidates',
-			array(
-				'methods'             => \WP_REST_Server::READABLE,
-				'callback'            => array( $this, 'list_candidates' ),
-				'permission_callback' => array( $this, 'check_admin_cap' ),
-				'args'                => array(
-					'search'       => array(
-						'type'              => 'string',
-						'sanitize_callback' => 'sanitize_text_field',
-					),
-					'cpf'          => array(
-						'type'              => 'string',
-						'sanitize_callback' => 'sanitize_text_field',
-					),
-					'rf'           => array(
-						'type'              => 'string',
-						'sanitize_callback' => 'sanitize_text_field',
-					),
-					'notice_id'    => array(
-						'type'              => 'integer',
-						'sanitize_callback' => 'absint',
-					),
-					'adjutancy_id' => array(
-						'type'              => 'integer',
-						'sanitize_callback' => 'absint',
-					),
-				),
-			)
-		);
-
-		register_rest_route(
-			$ns,
-			$base . '/candidates/(?P<id>\d+)',
-			array(
-				array(
-					'methods'             => \WP_REST_Server::READABLE,
-					'callback'            => array( $this, 'get_candidate' ),
-					'permission_callback' => array( $this, 'check_admin_cap' ),
-				),
-				array(
-					'methods'             => \WP_REST_Server::EDITABLE,
-					'callback'            => array( $this, 'update_candidate' ),
-					'permission_callback' => array( $this, 'check_admin_cap' ),
-				),
-				array(
-					'methods'             => \WP_REST_Server::DELETABLE,
-					'callback'            => array( $this, 'delete_candidate' ),
-					'permission_callback' => array( $this, 'check_admin_cap' ),
-				),
-			)
-		);
-
-		// ── Candidate-self dashboard ────────────────────────────────────
-		register_rest_route(
-			$ns,
-			$base . '/me/recruitment',
-			array(
-				'methods'             => \WP_REST_Server::READABLE,
-				'callback'            => array( $this, 'get_my_recruitment' ),
-				'permission_callback' => array( $this, 'check_logged_in' ),
-			)
-		);
-	}
-
-	// ─────────────────────────────────────────────────────────────────────
-	// Permission callbacks
-	// ─────────────────────────────────────────────────────────────────────
-
-	/**
-	 * Permission gate for every admin endpoint.
-	 *
-	 * The umbrella `ffc_manage_recruitment` cap remains the catch-all
-	 * — anyone holding it passes every admin route. The granular 6.2.0
-	 * caps (`ffc_view_recruitment`, `ffc_import_recruitment_csv`, etc.)
-	 * are layered on top via dedicated permission callbacks for the
-	 * higher-blast-radius routes ({@see self::check_can_view_recruitment()},
-	 * {@see self::check_can_import_csv()}, {@see self::check_can_call_candidates()}),
-	 * so an operator with only `ffc_view_recruitment` can hit GET endpoints
-	 * but not the CSV importer or call dispatcher.
-	 *
-	 * @return bool
-	 */
-	public function check_admin_cap(): bool {
-		return current_user_can( self::ADMIN_CAP );
-	}
-
-	/**
-	 * Permission gate for read-only endpoints. Accepts either the
-	 * granular `ffc_view_recruitment` cap or the umbrella
-	 * `ffc_manage_recruitment`.
-	 *
-	 * @since 6.2.0
-	 * @return bool
-	 */
-	public function check_can_view_recruitment(): bool {
-		return current_user_can( 'ffc_view_recruitment' ) || current_user_can( self::ADMIN_CAP );
-	}
-
-	/**
-	 * Permission gate for CSV import + promote-preview routes. The
-	 * highest-blast-radius operations on the module — replace entire
-	 * preview / definitive lists atomically. Accepts either the granular
-	 * `ffc_import_recruitment_csv` cap or the umbrella
-	 * `ffc_manage_recruitment`.
-	 *
-	 * @since 6.2.0
-	 * @return bool
-	 */
-	public function check_can_import_csv(): bool {
-		return current_user_can( 'ffc_import_recruitment_csv' ) || current_user_can( self::ADMIN_CAP );
-	}
-
-	/**
-	 * Permission gate for the call routes. Each call sends an email +
-	 * commits the candidate to a date / time, so operators that should
-	 * only manage data (without disparate communication authority) get
-	 * a separate cap. Accepts either the granular
-	 * `ffc_call_recruitment_candidates` cap or the umbrella
-	 * `ffc_manage_recruitment`.
-	 *
-	 * @since 6.2.0
-	 * @return bool
-	 */
-	public function check_can_call_candidates(): bool {
-		return current_user_can( 'ffc_call_recruitment_candidates' ) || current_user_can( self::ADMIN_CAP );
-	}
-
-	/**
-	 * Permission gate for the reasons-catalog routes. Reasons are global
-	 * across every notice — managing them is a config-style operation
-	 * separated from day-to-day notice management.
-	 *
-	 * @since 6.2.0
-	 * @return bool
-	 */
-	public function check_can_manage_reasons(): bool {
-		return current_user_can( 'ffc_manage_recruitment_reasons' ) || current_user_can( self::ADMIN_CAP );
-	}
-
-	/**
-	 * Permission gate for the candidate-self endpoint.
-	 *
-	 * @return bool
-	 */
-	public function check_logged_in(): bool {
-		return is_user_logged_in();
-	}
-
-	// ─────────────────────────────────────────────────────────────────────
-	// Notices
-	// ─────────────────────────────────────────────────────────────────────
-
-	/**
-	 * GET /notices
-	 *
-	 * @param \WP_REST_Request $request Request.
-	 * @return \WP_REST_Response
-	 */
-	public function list_notices( \WP_REST_Request $request ): \WP_REST_Response {
-		$status  = $request->get_param( 'status' );
-		$notices = RecruitmentNoticeRepository::get_all( is_string( $status ) && '' !== $status ? $status : null );
-
-		return new \WP_REST_Response( $notices, 200 );
-	}
-
-	/**
-	 * POST /notices
-	 *
-	 * @param \WP_REST_Request $request Request.
-	 * @return \WP_REST_Response|\WP_Error
-	 */
-	public function create_notice( \WP_REST_Request $request ) {
-		$code = (string) $request->get_param( 'code' );
-		$name = (string) $request->get_param( 'name' );
-
-		$id = RecruitmentNoticeRepository::create( $code, $name );
-		if ( false === $id ) {
-			return new \WP_Error(
-				'recruitment_notice_create_failed',
-				__( 'Failed to create notice (duplicate code?).', 'ffcertificate' ),
-				array( 'status' => 409 )
-			);
-		}
-
-		return new \WP_REST_Response( RecruitmentNoticeRepository::get_by_id( $id ), 201 );
-	}
-
-	/**
-	 * PATCH /notices/{id}
-	 *
-	 * @param \WP_REST_Request $request Request.
-	 * @return \WP_REST_Response|\WP_Error
-	 */
-	public function update_notice( \WP_REST_Request $request ) {
-		$id = (int) $request->get_param( 'id' );
-
-		// Status transitions go through the state machine so all gates and
-		// activity-log instrumentation fire correctly.
-		$new_status = $request->get_param( 'status' );
-		if ( is_string( $new_status ) && '' !== $new_status ) {
-			$reason     = $request->get_param( 'reason' );
-			$transition = RecruitmentNoticeStateMachine::transition_to(
-				$id,
-				$new_status,
-				is_string( $reason ) ? $reason : null
-			);
-			if ( ! $transition['success'] ) {
-				return $this->wp_error_from_envelope( $transition['errors'], 409 );
-			}
-		}
-
-		// Plain meta fields go through the repository update.
-		$meta = array_intersect_key(
-			$request->get_params(),
-			array_flip( array( 'name', 'code', 'public_columns_config' ) )
-		);
-		if ( ! empty( $meta ) ) {
-			$ok = RecruitmentNoticeRepository::update( $id, $meta );
-			if ( ! $ok ) {
-				return new \WP_Error(
-					'recruitment_notice_update_failed',
-					__( 'Notice update failed.', 'ffcertificate' ),
-					array( 'status' => 400 )
-				);
-			}
-		}
-
-		$notice = RecruitmentNoticeRepository::get_by_id( $id );
-		if ( null === $notice ) {
-			return new \WP_Error( 'recruitment_notice_not_found', RecruitmentErrorMessages::translate( 'recruitment_notice_not_found' ), array( 'status' => 404 ) );
-		}
-
-		return new \WP_REST_Response( $notice, 200 );
-	}
-
-	/**
-	 * GET /notices/{id}/classifications
-	 *
-	 * @param \WP_REST_Request $request Request.
-	 * @return \WP_REST_Response
-	 */
-	public function list_classifications( \WP_REST_Request $request ): \WP_REST_Response {
-		$id           = (int) $request->get_param( 'id' );
-		$list_type    = $request->get_param( 'list_type' );
-		$adjutancy_id = $request->get_param( 'adjutancy_id' );
-
-		$rows = RecruitmentClassificationRepository::get_for_notice(
-			$id,
-			is_string( $list_type ) && '' !== $list_type ? $list_type : null,
-			is_int( $adjutancy_id ) && $adjutancy_id > 0 ? $adjutancy_id : null
-		);
-
-		return new \WP_REST_Response( $rows, 200 );
-	}
-
-	/**
-	 * POST /notices/{id}/import
-	 *
-	 * @param \WP_REST_Request $request Request.
-	 * @return \WP_REST_Response|\WP_Error
-	 */
-	public function import_csv( \WP_REST_Request $request ) {
-		$id    = (int) $request->get_param( 'id' );
-		$files = $request->get_file_params();
-
-		if ( ! isset( $files['csv_file']['tmp_name'] ) || ! is_string( $files['csv_file']['tmp_name'] ) ) {
-			return new \WP_Error(
-				'recruitment_csv_file_missing',
-				__( 'CSV file is required.', 'ffcertificate' ),
-				array( 'status' => 400 )
-			);
-		}
-
-		// phpcs:ignore WordPress.WP.AlternativeFunctions.file_get_contents_file_get_contents -- Local upload tmp file; wp_remote_get is for HTTP only.
-		$content = file_get_contents( $files['csv_file']['tmp_name'] );
-		if ( false === $content ) {
-			return new \WP_Error(
-				'recruitment_csv_file_unreadable',
-				__( 'Could not read CSV file.', 'ffcertificate' ),
-				array( 'status' => 400 )
-			);
-		}
-
-		$result = RecruitmentCsvImporter::import_preview( $id, $content );
-
-		if ( ! $result['success'] ) {
-			return $this->wp_error_from_envelope( $result['errors'], 400 );
-		}
-
-		return new \WP_REST_Response( $result, 200 );
-	}
-
-	/**
-	 * POST /notices/{id}/promote-preview
-	 *
-	 * @param \WP_REST_Request $request Request.
-	 * @return \WP_REST_Response|\WP_Error
-	 */
-	public function promote_preview( \WP_REST_Request $request ) {
-		$id   = (int) $request->get_param( 'id' );
-		$mode = (string) $request->get_param( 'mode' );
-
-		if ( 'snapshot' === $mode ) {
-			$result = RecruitmentPromotionService::snapshot_to_definitive( $id );
-			if ( ! $result['success'] ) {
-				return $this->wp_error_from_envelope( $result['errors'], 409 );
-			}
-			return new \WP_REST_Response( $result, 200 );
-		}
-
-		// definitive_import mode requires an uploaded CSV.
-		$files = $request->get_file_params();
-		if ( ! isset( $files['csv_file']['tmp_name'] ) || ! is_string( $files['csv_file']['tmp_name'] ) ) {
-			return new \WP_Error(
-				'recruitment_csv_file_missing',
-				__( 'CSV file is required for definitive_import mode.', 'ffcertificate' ),
-				array( 'status' => 400 )
-			);
-		}
-
-		// phpcs:ignore WordPress.WP.AlternativeFunctions.file_get_contents_file_get_contents -- Local upload tmp file; wp_remote_get is for HTTP only.
-		$content = file_get_contents( $files['csv_file']['tmp_name'] );
-		if ( false === $content ) {
-			return new \WP_Error(
-				'recruitment_csv_file_unreadable',
-				__( 'Could not read CSV file.', 'ffcertificate' ),
-				array( 'status' => 400 )
-			);
-		}
-
-		$result = RecruitmentCsvImporter::import_definitive( $id, $content );
-		if ( ! $result['success'] ) {
-			return $this->wp_error_from_envelope( $result['errors'], 400 );
-		}
-
-		// Flip status to definitive after a successful definitive_import.
-		$transition = RecruitmentNoticeStateMachine::transition_to( $id, 'definitive' );
-		if ( ! $transition['success'] ) {
-			return $this->wp_error_from_envelope( $transition['errors'], 409 );
-		}
-
-		RecruitmentActivityLogger::notice_promoted( $id, 'definitive_import', 0 );
-
-		return new \WP_REST_Response( $result, 200 );
-	}
-
-	// ─────────────────────────────────────────────────────────────────────
-	// Classifications
-	// ─────────────────────────────────────────────────────────────────────
-
-	/**
-	 * POST /classifications/{id}/call
-	 *
-	 * @param \WP_REST_Request $request Request.
-	 * @return \WP_REST_Response|\WP_Error
-	 */
-	public function call_classification( \WP_REST_Request $request ) {
-		$id           = (int) $request->get_param( 'id' );
-		$reason_raw   = (string) ( $request->get_param( 'out_of_order_reason' ) ?? '' );
-		$notes_raw    = (string) ( $request->get_param( 'notes' ) ?? '' );
-		$reason_param = '' === $reason_raw ? null : $reason_raw;
-		$notes_param  = '' === $notes_raw ? null : $notes_raw;
-
-		$result = RecruitmentCallService::call_single(
-			$id,
-			(string) $request->get_param( 'date_to_assume' ),
-			(string) $request->get_param( 'time_to_assume' ),
-			get_current_user_id(),
-			$reason_param,
-			$notes_param
-		);
-
-		if ( ! $result['success'] ) {
-			return $this->wp_error_from_envelope( $result['errors'], 409 );
-		}
-
-		return new \WP_REST_Response( $result, 201 );
-	}
-
-	/**
-	 * POST /classifications/bulk-call
-	 *
-	 * @param \WP_REST_Request $request Request.
-	 * @return \WP_REST_Response|\WP_Error
-	 */
-	public function bulk_call_classifications( \WP_REST_Request $request ) {
-		$ids_raw = $request->get_param( 'classification_ids' );
-		if ( ! is_array( $ids_raw ) || empty( $ids_raw ) ) {
-			return new \WP_Error(
-				'recruitment_bulk_call_empty_id_list',
-				'',
-				array( 'status' => 400 )
-			);
-		}
-
-		$ids = array_values( array_filter( array_map( 'intval', $ids_raw ) ) );
-
-		$reasons_raw = $request->get_param( 'out_of_order_reasons' );
-		$reasons     = is_array( $reasons_raw ) ? array_map( 'sanitize_text_field', $reasons_raw ) : array();
-
-		$notes_raw   = (string) ( $request->get_param( 'notes' ) ?? '' );
-		$notes_param = '' === $notes_raw ? null : $notes_raw;
-
-		$result = RecruitmentCallService::call_bulk(
-			$ids,
-			(string) $request->get_param( 'date_to_assume' ),
-			(string) $request->get_param( 'time_to_assume' ),
-			get_current_user_id(),
-			$reasons,
-			$notes_param
-		);
-
-		if ( ! $result['success'] ) {
-			return $this->wp_error_from_envelope( $result['errors'], 409 );
-		}
-
-		return new \WP_REST_Response( $result, 201 );
-	}
-
-	/**
-	 * PATCH /classifications/{id}/status
-	 *
-	 * @param \WP_REST_Request $request Request.
-	 * @return \WP_REST_Response|\WP_Error
-	 */
-	public function change_classification_status( \WP_REST_Request $request ) {
-		$id     = (int) $request->get_param( 'id' );
-		$reason = $request->get_param( 'reason' );
-		$result = RecruitmentClassificationStateMachine::transition_to(
-			$id,
-			(string) $request->get_param( 'status' ),
-			is_string( $reason ) ? $reason : null
-		);
-		if ( ! $result['success'] ) {
-			return $this->wp_error_from_envelope( $result['errors'], 409 );
-		}
-
-		return new \WP_REST_Response( $result, 200 );
-	}
-
-	/**
-	 * PATCH /classifications/{id}/preview-status — set the preliminary
-	 * list's visual status (and optional reason). Visual-only — never
-	 * touches the §5.2 state machine on the definitive list.
-	 *
-	 * Validates:
-	 *   - the classification exists and is `list_type='preview'`
-	 *   - `preview_status` is a known enum value
-	 *   - the reason (if any) exists in the catalog AND its `applies_to`
-	 *     covers the requested status
-	 *   - if the per-status `preview_reason_required_*` flag is set in
-	 *     Settings, the reason must be supplied
-	 *
-	 * @param \WP_REST_Request $request Request.
-	 * @return \WP_REST_Response|\WP_Error
-	 */
-	public function change_classification_preview_status( \WP_REST_Request $request ) {
-		$id             = (int) $request->get_param( 'id' );
-		$preview_status = (string) $request->get_param( 'preview_status' );
-		$reason_id_raw  = $request->get_param( 'preview_reason_id' );
-		$reason_id      = is_numeric( $reason_id_raw ) ? (int) $reason_id_raw : 0;
-
-		$cls = RecruitmentClassificationRepository::get_by_id( $id );
-		if ( null === $cls ) {
-			return new \WP_Error( 'recruitment_classification_not_found', RecruitmentErrorMessages::translate( 'recruitment_classification_not_found' ), array( 'status' => 404 ) );
-		}
-		if ( 'preview' !== (string) $cls->list_type ) {
-			return new \WP_Error( 'recruitment_preview_status_only_on_preview_list', __( 'Preliminary status can only be set on classifications in the preview list.', 'ffcertificate' ), array( 'status' => 409 ) );
-		}
-
-		$valid_statuses = array( 'empty', 'denied', 'granted', 'appeal_denied', 'appeal_granted' );
-		if ( ! in_array( $preview_status, $valid_statuses, true ) ) {
-			return new \WP_Error( 'recruitment_preview_status_invalid', RecruitmentErrorMessages::translate( 'recruitment_preview_status_invalid' ), array( 'status' => 400 ) );
-		}
-
-		// 'empty' clears any pre-existing reason — operators reset by
-		// flipping the dropdown back to "no decision".
-		if ( 'empty' === $preview_status ) {
-			$reason_id = 0;
-		}
-
-		$settings = RecruitmentSettings::all();
-		if ( 'empty' !== $preview_status ) {
-			$required_key = 'preview_reason_required_' . $preview_status;
-			$required     = ! empty( $settings[ $required_key ] );
-			if ( $required && $reason_id <= 0 ) {
-				return new \WP_Error( 'recruitment_preview_reason_required', __( 'A reason is required for this preliminary status.', 'ffcertificate' ), array( 'status' => 400 ) );
-			}
-		}
-
-		if ( $reason_id > 0 ) {
-			$reason = RecruitmentReasonRepository::get_by_id( $reason_id );
-			if ( null === $reason ) {
-				return new \WP_Error( 'recruitment_preview_reason_not_found', RecruitmentErrorMessages::translate( 'recruitment_preview_reason_not_found' ), array( 'status' => 404 ) );
-			}
-			$applies = RecruitmentReasonRepository::decode_applies_to( (string) ( $reason->applies_to ?? '' ) );
-			if ( ! in_array( $preview_status, $applies, true ) ) {
-				return new \WP_Error( 'recruitment_preview_reason_status_mismatch', __( 'This reason cannot be used with the chosen preliminary status.', 'ffcertificate' ), array( 'status' => 400 ) );
-			}
-		}
-
-		$ok = RecruitmentClassificationRepository::set_preview_status( $id, $preview_status, $reason_id > 0 ? $reason_id : null );
-		if ( ! $ok ) {
-			return new \WP_Error( 'recruitment_preview_status_update_failed', RecruitmentErrorMessages::translate( 'recruitment_preview_status_update_failed' ), array( 'status' => 400 ) );
-		}
-
-		return new \WP_REST_Response( RecruitmentClassificationRepository::get_by_id( $id ), 200 );
-	}
-
-	/**
-	 * DELETE /classifications/{id}
-	 *
-	 * @param \WP_REST_Request $request Request.
-	 * @return \WP_REST_Response|\WP_Error
-	 */
-	public function delete_classification( \WP_REST_Request $request ) {
-		$id     = (int) $request->get_param( 'id' );
-		$result = RecruitmentDeleteService::delete_classification( $id );
-		if ( ! $result['success'] ) {
-			return $this->wp_error_from_envelope_with_blocked( $result, 409 );
-		}
-		return new \WP_REST_Response( $result, 200 );
-	}
-
-	/**
-	 * POST /classifications/{id}/calls/{call_id}/cancel
-	 *
-	 * @param \WP_REST_Request $request Request.
-	 * @return \WP_REST_Response|\WP_Error
-	 */
-	public function cancel_call( \WP_REST_Request $request ) {
-		$call_id = (int) $request->get_param( 'call_id' );
-		$result  = RecruitmentCallService::cancel_call(
-			$call_id,
-			(string) $request->get_param( 'reason' ),
-			get_current_user_id()
-		);
-		if ( ! $result['success'] ) {
-			return $this->wp_error_from_envelope( $result['errors'], 409 );
-		}
-		return new \WP_REST_Response( $result, 200 );
-	}
-
-	// ─────────────────────────────────────────────────────────────────────
-	// Adjutancies
-	// ─────────────────────────────────────────────────────────────────────
-
-	/**
-	 * GET /adjutancies — list every adjutancy.
-	 *
-	 * @return \WP_REST_Response
-	 */
-	public function list_adjutancies(): \WP_REST_Response {
-		return new \WP_REST_Response( RecruitmentAdjutancyRepository::get_all(), 200 );
-	}
-
-	/**
-	 * POST /adjutancies — create a new adjutancy.
-	 *
-	 * @param \WP_REST_Request $request Request.
-	 * @return \WP_REST_Response|\WP_Error
-	 */
-	public function create_adjutancy( \WP_REST_Request $request ) {
-		$id = RecruitmentAdjutancyRepository::create(
-			(string) $request->get_param( 'slug' ),
-			(string) $request->get_param( 'name' ),
-			(string) ( $request->get_param( 'color' ) ?? '' )
-		);
-		if ( false === $id ) {
-			return new \WP_Error(
-				'recruitment_adjutancy_create_failed',
-				__( 'Adjutancy creation failed (duplicate slug?).', 'ffcertificate' ),
-				array( 'status' => 409 )
-			);
-		}
-		return new \WP_REST_Response( RecruitmentAdjutancyRepository::get_by_id( $id ), 201 );
-	}
-
-	/**
-	 * PATCH /adjutancies/{id} — update slug or name.
-	 *
-	 * @param \WP_REST_Request $request Request.
-	 * @return \WP_REST_Response|\WP_Error
-	 */
-	public function update_adjutancy( \WP_REST_Request $request ) {
-		$id   = (int) $request->get_param( 'id' );
-		$data = array_intersect_key( $request->get_params(), array_flip( array( 'slug', 'name', 'color' ) ) );
-		$ok   = RecruitmentAdjutancyRepository::update( $id, $data );
-		if ( ! $ok ) {
-			return new \WP_Error(
-				'recruitment_adjutancy_update_failed',
-				'',
-				array( 'status' => 400 )
-			);
-		}
-		return new \WP_REST_Response( RecruitmentAdjutancyRepository::get_by_id( $id ), 200 );
-	}
-
-	/**
-	 * DELETE /adjutancies/{id} — gated removal.
-	 *
-	 * @param \WP_REST_Request $request Request.
-	 * @return \WP_REST_Response|\WP_Error
-	 */
-	public function delete_adjutancy( \WP_REST_Request $request ) {
-		$result = RecruitmentDeleteService::delete_adjutancy( (int) $request->get_param( 'id' ) );
-		if ( ! $result['success'] ) {
-			return $this->wp_error_from_envelope_with_blocked( $result, 409 );
-		}
-		return new \WP_REST_Response( $result, 200 );
-	}
-
-	// ─────────────────────────────────────────────────────────────────────
-	// Reasons
-	// ─────────────────────────────────────────────────────────────────────
-
-	/**
-	 * GET /reasons — list every reason in the global catalog.
-	 *
-	 * @return \WP_REST_Response
-	 */
-	public function list_reasons(): \WP_REST_Response {
-		return new \WP_REST_Response( RecruitmentReasonRepository::get_all(), 200 );
-	}
-
-	/**
-	 * POST /reasons — create a new reason.
-	 *
-	 * @param \WP_REST_Request $request Request.
-	 * @return \WP_REST_Response|\WP_Error
-	 */
-	public function create_reason( \WP_REST_Request $request ) {
-		$applies_raw = $request->get_param( 'applies_to' );
-		$applies     = array();
-		if ( is_array( $applies_raw ) ) {
-			foreach ( $applies_raw as $candidate ) {
-				if ( is_string( $candidate ) ) {
-					$applies[] = $candidate;
-				}
-			}
-		}
-
-		$id = RecruitmentReasonRepository::create(
-			(string) $request->get_param( 'slug' ),
-			(string) $request->get_param( 'label' ),
-			(string) ( $request->get_param( 'color' ) ?? '' ),
-			$applies
-		);
-		if ( false === $id ) {
-			return new \WP_Error(
-				'recruitment_reason_create_failed',
-				__( 'Reason creation failed (duplicate slug?).', 'ffcertificate' ),
-				array( 'status' => 409 )
-			);
-		}
-		return new \WP_REST_Response( RecruitmentReasonRepository::get_by_id( $id ), 201 );
-	}
-
-	/**
-	 * PATCH /reasons/{id} — update slug / label / color / applies_to.
-	 *
-	 * @param \WP_REST_Request $request Request.
-	 * @return \WP_REST_Response|\WP_Error
-	 */
-	public function update_reason( \WP_REST_Request $request ) {
-		$id   = (int) $request->get_param( 'id' );
-		$data = array_intersect_key( $request->get_params(), array_flip( array( 'slug', 'label', 'color', 'applies_to' ) ) );
-		$ok   = RecruitmentReasonRepository::update( $id, $data );
-		if ( ! $ok ) {
-			return new \WP_Error(
-				'recruitment_reason_update_failed',
-				'',
-				array( 'status' => 400 )
-			);
-		}
-		return new \WP_REST_Response( RecruitmentReasonRepository::get_by_id( $id ), 200 );
-	}
-
-	/**
-	 * DELETE /reasons/{id} — gated by the references count.
-	 *
-	 * @param \WP_REST_Request $request Request.
-	 * @return \WP_REST_Response|\WP_Error
-	 */
-	public function delete_reason( \WP_REST_Request $request ) {
-		$id    = (int) $request->get_param( 'id' );
-		$count = RecruitmentReasonRepository::count_references( $id );
-		if ( $count > 0 ) {
-			return new \WP_Error(
-				'recruitment_reason_in_use',
-				__( 'Cannot delete: this reason is still attached to at least one classification.', 'ffcertificate' ),
-				array(
-					'status'          => 409,
-					'reference_count' => $count,
-				)
-			);
-		}
-		$ok = RecruitmentReasonRepository::delete( $id );
-		if ( ! $ok ) {
-			return new \WP_Error( 'recruitment_reason_delete_failed', '', array( 'status' => 400 ) );
-		}
-		return new \WP_REST_Response( array( 'success' => true ), 200 );
-	}
-
-	/**
-	 * PUT /notices/{id}/adjutancies/{adjutancy_id} — attach the adjutancy
-	 * to the notice via the N:N junction. Idempotent: re-attaching an
-	 * already-attached pair returns 200 with `created=false`.
-	 *
-	 * Both ids must reference existing rows; otherwise returns 404.
-	 *
-	 * @param \WP_REST_Request $request Request.
-	 * @return \WP_REST_Response|\WP_Error
-	 */
-	public function attach_notice_adjutancy( \WP_REST_Request $request ) {
-		$notice_id    = (int) $request->get_param( 'id' );
-		$adjutancy_id = (int) $request->get_param( 'adjutancy_id' );
-
-		if ( null === RecruitmentNoticeRepository::get_by_id( $notice_id ) ) {
-			return new \WP_Error( 'recruitment_notice_not_found', 'Notice not found.', array( 'status' => 404 ) );
-		}
-		if ( null === RecruitmentAdjutancyRepository::get_by_id( $adjutancy_id ) ) {
-			return new \WP_Error( 'recruitment_adjutancy_not_found', 'Adjutancy not found.', array( 'status' => 404 ) );
-		}
-
-		$already = RecruitmentNoticeAdjutancyRepository::is_attached( $notice_id, $adjutancy_id );
-		if ( $already ) {
-			return new \WP_REST_Response(
-				array(
-					'notice_id'    => $notice_id,
-					'adjutancy_id' => $adjutancy_id,
-					'created'      => false,
-				),
-				200
-			);
-		}
-
-		$ok = RecruitmentNoticeAdjutancyRepository::attach( $notice_id, $adjutancy_id );
-		if ( ! $ok ) {
-			return new \WP_Error(
-				'recruitment_notice_adjutancy_attach_failed',
-				'Failed to attach adjutancy to notice.',
-				array( 'status' => 500 )
-			);
-		}
-
-		return new \WP_REST_Response(
-			array(
-				'notice_id'    => $notice_id,
-				'adjutancy_id' => $adjutancy_id,
-				'created'      => true,
-			),
-			201
-		);
-	}
-
-	/**
-	 * DELETE /notices/{id}/adjutancies/{adjutancy_id} — detach the
-	 * adjutancy from the notice. Idempotent: detaching a pair that's
-	 * already absent returns 200.
-	 *
-	 * @param \WP_REST_Request $request Request.
-	 * @return \WP_REST_Response
-	 */
-	public function detach_notice_adjutancy( \WP_REST_Request $request ): \WP_REST_Response {
-		$notice_id    = (int) $request->get_param( 'id' );
-		$adjutancy_id = (int) $request->get_param( 'adjutancy_id' );
-
-		RecruitmentNoticeAdjutancyRepository::detach( $notice_id, $adjutancy_id );
-
-		return new \WP_REST_Response(
-			array(
-				'notice_id'    => $notice_id,
-				'adjutancy_id' => $adjutancy_id,
-				'attached'     => false,
-			),
-			200
-		);
-	}
-
-	// ─────────────────────────────────────────────────────────────────────
-	// Candidates
-	// ─────────────────────────────────────────────────────────────────────
-
-	/**
-	 * GET /candidates — filter by cpf/rf/name/notice/adjutancy.
-	 *
-	 * @param \WP_REST_Request $request Request.
-	 * @return \WP_REST_Response|\WP_Error
-	 */
-	public function list_candidates( \WP_REST_Request $request ) {
-		// CPF/RF are passed as plaintext digits; hash here for the lookup.
-		$cpf = $request->get_param( 'cpf' );
-		if ( is_string( $cpf ) && '' !== $cpf ) {
-			$cpf_digits = preg_replace( '/[^0-9]/', '', $cpf ) ?? '';
-			if ( '' !== $cpf_digits ) {
-				$candidate = RecruitmentCandidateRepository::get_by_cpf_hash( (string) Encryption::hash( $cpf_digits ) );
-				return new \WP_REST_Response( null === $candidate ? array() : array( $this->shape_candidate_admin( $candidate ) ), 200 );
-			}
-		}
-
-		$rf = $request->get_param( 'rf' );
-		if ( is_string( $rf ) && '' !== $rf ) {
-			$rf_digits = preg_replace( '/[^0-9]/', '', $rf ) ?? '';
-			if ( '' !== $rf_digits ) {
-				$candidate = RecruitmentCandidateRepository::get_by_rf_hash( (string) Encryption::hash( $rf_digits ) );
-				return new \WP_REST_Response( null === $candidate ? array() : array( $this->shape_candidate_admin( $candidate ) ), 200 );
-			}
-		}
-
-		// Generic listing not yet implemented in repository (out of scope for
-		// sprint 9.1 MVP — search by name will land alongside the admin UI).
-		return new \WP_Error(
-			'recruitment_candidate_list_requires_filter',
-			__( 'Candidate listing requires a cpf or rf filter for now.', 'ffcertificate' ),
-			array( 'status' => 400 )
-		);
-	}
-
-	/**
-	 * GET /candidates/{id} — admin view with decrypted fields.
-	 *
-	 * @param \WP_REST_Request $request Request.
-	 * @return \WP_REST_Response|\WP_Error
-	 */
-	public function get_candidate( \WP_REST_Request $request ) {
-		$id        = (int) $request->get_param( 'id' );
-		$candidate = RecruitmentCandidateRepository::get_by_id( $id );
-		if ( null === $candidate ) {
-			return new \WP_Error( 'recruitment_candidate_not_found', '', array( 'status' => 404 ) );
-		}
-		return new \WP_REST_Response( $this->shape_candidate_admin( $candidate ), 200 );
-	}
-
-	/**
-	 * PATCH /candidates/{id} — re-encrypts on cpf/rf/email change.
-	 *
-	 * @param \WP_REST_Request $request Request.
-	 * @return \WP_REST_Response|\WP_Error
-	 */
-	public function update_candidate( \WP_REST_Request $request ) {
-		$id   = (int) $request->get_param( 'id' );
-		$data = $request->get_params();
-
-		$update = array_intersect_key( $data, array_flip( array( 'name', 'phone', 'notes' ) ) );
-
-		// Encrypt cpf/rf/email via the SensitiveFieldRegistry if supplied.
-		$plaintexts = array();
-		foreach ( array( 'cpf', 'rf', 'email' ) as $key ) {
-			if ( isset( $data[ $key ] ) && is_string( $data[ $key ] ) && '' !== trim( $data[ $key ] ) ) {
-				$value = trim( $data[ $key ] );
-				if ( 'email' === $key ) {
-					$value = strtolower( $value );
-				} else {
-					$value = preg_replace( '/[^0-9]/', '', $value ) ?? '';
-				}
-				$plaintexts[ $key ] = $value;
-			}
-		}
-
-		if ( ! empty( $plaintexts ) ) {
-			$encrypted = SensitiveFieldRegistry::encrypt_fields(
-				SensitiveFieldRegistry::CONTEXT_RECRUITMENT_CANDIDATE,
-				$plaintexts
-			);
-			$update    = array_merge( $update, $encrypted );
-		}
-
-		if ( empty( $update ) ) {
-			return new \WP_Error(
-				'recruitment_candidate_update_no_writable_fields',
-				'',
-				array( 'status' => 400 )
-			);
-		}
-
-		$ok = RecruitmentCandidateRepository::update( $id, $update );
-		if ( ! $ok ) {
-			return new \WP_Error(
-				'recruitment_candidate_update_failed',
-				'',
-				array( 'status' => 409 )
-			);
-		}
-
-		$candidate = RecruitmentCandidateRepository::get_by_id( $id );
-		return new \WP_REST_Response(
-			null === $candidate ? null : $this->shape_candidate_admin( $candidate ),
-			200
-		);
-	}
-
-	/**
-	 * DELETE /candidates/{id} — gated hard-delete (zero classifications).
-	 *
-	 * @param \WP_REST_Request $request Request.
-	 * @return \WP_REST_Response|\WP_Error
-	 */
-	public function delete_candidate( \WP_REST_Request $request ) {
-		$result = RecruitmentDeleteService::delete_candidate( (int) $request->get_param( 'id' ) );
-		if ( ! $result['success'] ) {
-			return $this->wp_error_from_envelope_with_blocked( $result, 409 );
-		}
-		return new \WP_REST_Response( $result, 200 );
-	}
-
-	// ─────────────────────────────────────────────────────────────────────
-	// Candidate-self dashboard
-	// ─────────────────────────────────────────────────────────────────────
-
-	/**
-	 * GET /me/recruitment.
-	 *
-	 * @return \WP_REST_Response
-	 */
-	public function get_my_recruitment(): \WP_REST_Response {
-		$user_id = get_current_user_id();
-		if ( $user_id <= 0 ) {
-			return new \WP_REST_Response( array(), 200 );
-		}
-
-		$candidates = RecruitmentCandidateRepository::get_by_user_id( $user_id );
-		if ( empty( $candidates ) ) {
-			return new \WP_REST_Response( array(), 200 );
-		}
-
-		// Group classifications by notice for the §9.2 grouped payload.
-		$out = array();
-		foreach ( $candidates as $candidate ) {
-			$candidate_id    = (int) $candidate->id;
-			$classifications = RecruitmentClassificationRepository::get_for_candidate( $candidate_id );
-
-			foreach ( $classifications as $cls ) {
-				$notice_id = (int) $cls->notice_id;
-				$notice    = RecruitmentNoticeRepository::get_by_id( $notice_id );
-				if ( null === $notice || 'draft' === $notice->status ) {
-					continue; // Draft notices are never exposed.
-				}
-
-				if ( ! isset( $out[ $notice_id ] ) ) {
-					$out[ $notice_id ] = array(
-						'notice'          => array(
-							'id'           => $notice_id,
-							'code'         => $notice->code,
-							'name'         => $notice->name,
-							'status'       => $notice->status,
-							'was_reopened' => '1' === $notice->was_reopened,
-						),
-						'classifications' => array(),
-					);
-				}
-
-				$out[ $notice_id ]['classifications'][] = array(
-					'id'        => (int) $cls->id,
-					'list_type' => $cls->list_type,
-					'rank'      => (int) $cls->rank,
-					'score'     => $cls->score,
-					'status'    => $cls->status,
-					'calls'     => RecruitmentCallRepository::get_history_for_classification( (int) $cls->id ),
-				);
-			}
-		}
-
-		return new \WP_REST_Response( array_values( $out ), 200 );
-	}
-
-	// ─────────────────────────────────────────────────────────────────────
-	// Helpers
-	// ─────────────────────────────────────────────────────────────────────
-
-	/**
-	 * Decrypt sensitive fields for admin-side candidate responses.
-	 *
-	 * @param object $candidate Candidate row (CandidateRow shape).
-	 * @phpstan-param CandidateRow $candidate
-	 * @return array<string, mixed>
-	 */
-	private function shape_candidate_admin( object $candidate ): array {
-		$decrypt = static function ( $value ) {
-			if ( ! is_string( $value ) || '' === $value ) {
-				return null;
-			}
-			$plain = Encryption::decrypt( $value );
-			return null === $plain ? null : $plain;
-		};
-
-		$mask_or_plain_email = function ( ?string $plain ): ?string {
-			return null === $plain ? null : DocumentFormatter::mask_email( $plain );
-		};
-
-		return array(
-			'id'           => (int) $candidate->id,
-			'user_id'      => null === $candidate->user_id ? null : (int) $candidate->user_id,
-			'name'         => $candidate->name,
-			'cpf'          => $decrypt( $candidate->cpf_encrypted ),
-			'rf'           => $decrypt( $candidate->rf_encrypted ),
-			'email'        => $decrypt( $candidate->email_encrypted ),
-			'email_masked' => $mask_or_plain_email( $decrypt( $candidate->email_encrypted ) ),
-			'phone'        => $candidate->phone,
-			'notes'        => $candidate->notes,
-			'is_pcd'       => RecruitmentPcdHasher::verify( $candidate->pcd_hash, (int) $candidate->id ),
-			'created_at'   => $candidate->created_at,
-			'updated_at'   => $candidate->updated_at,
-		);
-	}
-
-	/**
-	 * Convert an envelope-style errors array into a `\WP_Error`.
-	 *
-	 * @param array<int, string> $errors  Error codes.
-	 * @param int                $status  HTTP status code.
-	 * @return \WP_Error
-	 */
-	private function wp_error_from_envelope( array $errors, int $status ): \WP_Error {
-		$code    = $errors[0] ?? 'recruitment_error';
-		$message = RecruitmentErrorMessages::translate( $code );
-		return new \WP_Error(
-			$code,
-			$message,
-			array(
-				'status'   => $status,
-				'errors'   => $errors,
-				'messages' => RecruitmentErrorMessages::translate_all( $errors ),
-			)
-		);
-	}
-
-	/**
-	 * Same as `wp_error_from_envelope` but additionally surfaces a
-	 * `blocked_by` reference-count map (returned by the delete service).
-	 *
-	 * @param array{success: bool, errors: list<string>, blocked_by?: array<string, int>} $envelope Service-result envelope.
-	 * @param int                                                                         $status   HTTP status code.
-	 * @return \WP_Error
-	 */
-	private function wp_error_from_envelope_with_blocked( array $envelope, int $status ): \WP_Error {
-		$first   = $envelope['errors'][0] ?? 'recruitment_error';
-		$message = RecruitmentErrorMessages::translate( $first );
-		$data    = array(
-			'status'   => $status,
-			'errors'   => $envelope['errors'],
-			'messages' => RecruitmentErrorMessages::translate_all( $envelope['errors'] ),
-		);
-		if ( isset( $envelope['blocked_by'] ) ) {
-			$data['blocked_by'] = $envelope['blocked_by'];
-		}
-		return new \WP_Error( $first, $message, $data );
-	}
-}
