+++ tests/Unit/RecruitmentRestControllerTest.php
@@ -8,17 +8,31 @@ use Brain\Monkey\Functions;
 use Mockery;
 use Mockery\Adapter\Phpunit\MockeryPHPUnitIntegration;
 use PHPUnit\Framework\TestCase;
-use FreeFormCertificate\Recruitment\RecruitmentRestController;
+use FreeFormCertificate\Recruitment\RecruitmentNoticesRestController;
+use FreeFormCertificate\Recruitment\RecruitmentClassificationsRestController;
+use FreeFormCertificate\Recruitment\RecruitmentAdjutanciesRestController;
+use FreeFormCertificate\Recruitment\RecruitmentCandidatesRestController;
 
 /**
- * Tests for RecruitmentRestController — focused on permission gating and
+ * Tests for the recruitment REST surface — focused on permission gating and
  * route registration. Endpoint dispatch logic is exercised indirectly by
  * the service-layer tests (CallService, StateMachine, DeleteService) which
- * cover every error code the controller surfaces. The controller is a thin
- * adapter; full end-to-end integration tests for the REST surface land in
+ * cover every error code the controllers surface. The controllers are thin
+ * adapters; full end-to-end integration tests for the REST surface land in
  * sprint 13.
  *
- * @covers \FreeFormCertificate\Recruitment\RecruitmentRestController
+ * After sprint S2 of #141 the original god-object `RecruitmentRestController`
+ * was split into four domain controllers sharing a common
+ * `RecruitmentRestSupport` trait. The cap-check tests instantiate a single
+ * controller (the trait methods are shared by all four); the
+ * route-registration tests instantiate every controller and accumulate the
+ * registrations into a shared array so the existing spot-checks against
+ * the full route set still hold.
+ *
+ * @covers \FreeFormCertificate\Recruitment\RecruitmentNoticesRestController
+ * @covers \FreeFormCertificate\Recruitment\RecruitmentClassificationsRestController
+ * @covers \FreeFormCertificate\Recruitment\RecruitmentAdjutanciesRestController
+ * @covers \FreeFormCertificate\Recruitment\RecruitmentCandidatesRestController
  */
 class RecruitmentRestControllerTest extends TestCase {
 
@@ -45,7 +59,7 @@ class RecruitmentRestControllerTest extends TestCase {
 			}
 		);
 
-		$controller = new RecruitmentRestController();
+		$controller = new RecruitmentNoticesRestController();
 		$this->assertTrue( $controller->check_admin_cap() );
 		$this->assertSame( 'ffc_manage_recruitment', $captured_cap );
 	}
@@ -53,21 +67,21 @@ class RecruitmentRestControllerTest extends TestCase {
 	public function test_check_admin_cap_returns_false_without_capability(): void {
 		Functions\when( 'current_user_can' )->justReturn( false );
 
-		$controller = new RecruitmentRestController();
+		$controller = new RecruitmentNoticesRestController();
 		$this->assertFalse( $controller->check_admin_cap() );
 	}
 
 	public function test_check_logged_in_returns_true_when_authenticated(): void {
 		Functions\when( 'is_user_logged_in' )->justReturn( true );
 
-		$controller = new RecruitmentRestController();
+		$controller = new RecruitmentNoticesRestController();
 		$this->assertTrue( $controller->check_logged_in() );
 	}
 
 	public function test_check_logged_in_returns_false_for_anonymous(): void {
 		Functions\when( 'is_user_logged_in' )->justReturn( false );
 
-		$controller = new RecruitmentRestController();
+		$controller = new RecruitmentNoticesRestController();
 		$this->assertFalse( $controller->check_logged_in() );
 	}
 
@@ -83,8 +97,10 @@ class RecruitmentRestControllerTest extends TestCase {
 			}
 		);
 
-		$controller = new RecruitmentRestController();
-		$controller->register_routes();
+		( new RecruitmentNoticesRestController() )->register_routes();
+		( new RecruitmentClassificationsRestController() )->register_routes();
+		( new RecruitmentAdjutanciesRestController() )->register_routes();
+		( new RecruitmentCandidatesRestController() )->register_routes();
 
 		$this->assertNotEmpty( $registered, 'register_routes() must call register_rest_route at least once' );
 
@@ -103,8 +119,10 @@ class RecruitmentRestControllerTest extends TestCase {
 			}
 		);
 
-		$controller = new RecruitmentRestController();
-		$controller->register_routes();
+		( new RecruitmentNoticesRestController() )->register_routes();
+		( new RecruitmentClassificationsRestController() )->register_routes();
+		( new RecruitmentAdjutanciesRestController() )->register_routes();
+		( new RecruitmentCandidatesRestController() )->register_routes();
 
 		// Spot-check the major admin routes documented in §14 of the plan.
 		$this->assertContains( '/recruitment/notices', $registered );
